Change naming of within asserts
Change the name of within asserts to match the naming convention
of other asserts.

Previously the words "EQUAL" and "NOT_EQUAL" of those asserts
used to be in the middle, they're now changed to be at the end of
the assert's name to be consistent with the names of all other asserts.