Refactor how we handle dynamic arrays
Instead of mannually creating an array, an append and pop function
for each type we want an array of,
we made a macro-based dynamic array creator that automatically declares
an array of a type and has an append, pop, and init macro
written only once.

Also, relocated allocate() and reallocate() to allocator.c/.h
so dynamic_array.h can use them.